Backend folder modularization
