+++ desktop/src-tauri/src/api/mod.rs
@@ -1,0 +1,3 @@
+pub mod commands;
+pub mod events;
+pub mod tcpservice;
