+++ desktop/src-tauri/src/lib.rs
@@ -1,26 +1,25 @@
 // Learn more about Tauri commands at https://tauri.app/develop/calling-rust/
+mod api;
+use api::commands::commands;
+use api::tcpservice::tcpservice;
+
 use serde::Serialize;
-use serde_json;
-use tokio::{io::AsyncWriteExt, net::TcpListener};
-use tauri::{Emitter, Manager, AppHandle};
+use tauri::Manager;
 
 use lazy_static::lazy_static;
 use serde::Deserialize;
-use std::{
-    sync::{Arc, Mutex},
-    time::Duration,
-};
-use local_ip_address::local_ip;
+use std::sync::{Arc, Mutex};
+
 
 #[cfg_attr(mobile, tauri::mobile_entry_point)]
 pub fn run() {
     tauri::Builder::default()
         .plugin(tauri_plugin_opener::init())
-        .invoke_handler(tauri::generate_handler![update_slider, get_ip_address])
+        .invoke_handler(tauri::generate_handler![commands::update_slider, commands::get_ip_address])
         .setup(|app| {
 
             let app_handle = app.app_handle().clone();
-            tauri::async_runtime::spawn(tcp_listener(app_handle));
+            tauri::async_runtime::spawn(tcpservice::tcp_listener(app_handle));
             Ok(())
         })
         .run(tauri::generate_context!())
@@ -28,7 +27,7 @@ pub fn run() {
 }
 
 lazy_static! {
-    static ref PARAMS: Arc<Mutex<Parameters>> = Arc::new(Mutex::new(Parameters {
+    pub static ref PARAMS: Arc<Mutex<Parameters>> = Arc::new(Mutex::new(Parameters {
         slider1: 1.0,
         slider2: 1.0,
         slider3: 1.0,
@@ -37,81 +36,9 @@ lazy_static! {
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
-struct Parameters {
+pub struct Parameters {
     slider1: f32,
     slider2: f32,
     slider3: f32,
     slider4: f32,
-}
-
-async fn tcp_listener(app_handle: AppHandle) {
-    let addr = "0.0.0.0:9001"; // Listen on all interfaces
-    let listener = TcpListener::bind(&addr).await.expect("Failed to bind");
-
-    println!("WebSocket server listening on ws://{}", addr);
-
-    // let (tx, _) = broadcast::channel::<String>(10); // Message broadcaster
-
-    while let Ok((stream, addr)) = listener.accept().await {
-        println!("New connection: {}", addr);
-        let app_handle_clone = app_handle.clone();
-        tokio::spawn(handle_connection(stream, addr, app_handle_clone));
-    }
-}
-
-async fn handle_connection(mut stream: tokio::net::TcpStream, addr: std::net::SocketAddr, app_handle: AppHandle) {
-    
-    //Emit to frontend that device has connected
-    emit_connection_event(&app_handle);
-
-    loop {
-        let params = PARAMS.lock().unwrap().clone();
-        let json_string = serde_json::to_string(&params).unwrap();
-
-        // Breaks out of the loop when the Tcp client has disconnected and can not recieve more writes
-        if let Err(_) = stream.write_all((json_string + "\n").as_bytes()).await {
-            println!("Client disconnected: {}", addr);
-            emit_disconnection_event(app_handle);
-            break;
-        }
-        tokio::time::sleep(Duration::from_millis(1)).await;
-    }
-}
-
-/// Function called by Tauri UI when a slider value changes
-#[tauri::command]
-fn update_slider(slider_number: &str, slider_value: f32) {
-    let mut params = PARAMS.lock().unwrap();
-    match slider_number {
-        "1" => params.slider1 = slider_value,
-        "2" => params.slider2 = slider_value,
-        "3" => params.slider3 = slider_value,
-        "4" => params.slider4 = slider_value,
-        _ => {}
-    }
-    println!("Updated slider values {:?}", params);
-}
-
-#[tauri::command]
-fn get_ip_address() -> String {
-    match local_ip() {
-        Ok(ip) => {
-            println!("Local IP address: {}", ip);
-            return format!("{}", ip);
-        },
-        Err(e) => {
-            eprintln!("Error retrieving local IP: {}", e);
-            return String::from("");
-        },
-    }
-}
-
-// #[tauri::command]
-fn emit_connection_event(app_handle: &AppHandle) {
-    app_handle.emit("connection", "A device has connected").unwrap();
-}
-
-// #[tauri::command]
-fn emit_disconnection_event(app_handle: AppHandle) {
-    app_handle.emit("disconnection", "A device has disconnected").unwrap();
 }
